pyproject.toml
[tool.poetry]
name = "acelerado"
version = "0.1.0"
description = ""
authors = ["Waine Oliveira Jr <waine.oliveira.jr@gmail.com>"]
readme = "README.md"

classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
]

packages = [{include = "acelerado"}]

[tool.poetry.scripts]
acelerado = "acelerado.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.3.2"
google-api-python-client = "^2.135.0"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
pydantic = "^2.8.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
